requirements_demo.txt
torch==1.10.2 
torchtext==0.11.2
torchvision==0.11.3
tensorflow
dominate==2.4.0
dill
scipy
pandas
Pillow==7.0.0
pyyaml
scikit-image
imgaug
sklearn
tqdm
natsort
wandb 
